green: fail_when_try_to_delete_a_topic_without_topic_id_assigned

+++ Papyrus.Business/Topics/TopicService.cs
@@ -38,6 +38,10 @@ namespace Papyrus.Business.Topics
 
         public async Task Delete(Topic topic)
         {
+            if (string.IsNullOrWhiteSpace(topic.TopicId))
+            {
+                throw new CannotDeleteTopicsWithoutTopicIdAssignedException();
+            }
             await TopicRepository.Delete(topic);
         }
